fix: wrap remaining np.mean in float() for numpy-2.x mypy parity
CI installs numpy 2.x whose stricter type stubs return floating[Any] (not
float), failing the typed dict at disagreement.py:256 on 3.10/3.11 only (the
3.9 floor uses older numpy with lenient stubs and passed). Wraps the last
key_features entry in float() to match the adjacent entries; value-identical.

+++ src/evaluation/disagreement.py
@@ -253,8 +253,13 @@ class DisagreementAnalyzer:
                             np.mean([p.spread_magnitude for p in preds])
                         ),
                         "upset_rate": float(np.mean([p.y_true for p in preds])),
-                        f"avg_{prob_attr.replace('_prob', '')}_confidence": np.mean(
-                            [abs(getattr(p, prob_attr) - self.threshold) for p in preds]
+                        f"avg_{prob_attr.replace('_prob', '')}_confidence": float(
+                            np.mean(
+                                [
+                                    abs(getattr(p, prob_attr) - self.threshold)
+                                    for p in preds
+                                ]
+                            )
                         ),
                     },
                     example_games=[p.game_id for p in preds[:5]],
